refactor: 정기 모임 참여 비지니스 로직 분리
- 정기 모임 service 복잡성 해소
