+++ src/main/java/lems/cowshed/api/controller/regular/event/participation/RegularEventParticipationController.java
@@ -3,6 +3,7 @@ package lems.cowshed.api.controller.regular.event.participation;
 import lems.cowshed.api.controller.CommonResponse;
 import lems.cowshed.api.controller.dto.regular.event.response.RegularParticipantsInfo;
 import lems.cowshed.domain.user.CustomUserDetails;
+import lems.cowshed.service.regular.event.RegularEventParticipationService;
 import lems.cowshed.service.regular.event.RegularEventService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -13,25 +14,25 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class RegularEventParticipationController implements RegularEventParticipationSpecification {
 
-    private final RegularEventService regularEventService;
+    private final RegularEventParticipationService participationService;
 
     @PostMapping("/{regular-id}/participation")
     public CommonResponse<Void> save(@PathVariable("regular-id") Long regularId,
                                      @AuthenticationPrincipal CustomUserDetails userDetails){
-        regularEventService.saveParticipation(regularId,userDetails.getUserId());
+        participationService.saveParticipation(regularId,userDetails.getUserId());
         return CommonResponse.success();
     }
 
     @GetMapping("/{regular-id}/participants")
     public CommonResponse<RegularParticipantsInfo> getRegularParticipants(@PathVariable("regular-id") Long regularId){
-        RegularParticipantsInfo participantsInfo = regularEventService.getRegularParticipants(regularId);
+        RegularParticipantsInfo participantsInfo = participationService.getRegularParticipants(regularId);
         return CommonResponse.success(participantsInfo);
     }
 
     @DeleteMapping("/participation/{participation-id}")
     public CommonResponse<Void> delete(@PathVariable("participation-id") Long participationId,
                                        @AuthenticationPrincipal CustomUserDetails userDetails) {
-        regularEventService.deleteParticipation(participationId, userDetails.getUserId());
+        participationService.deleteParticipation(participationId, userDetails.getUserId());
         return CommonResponse.success();
     }
 }
