+++ src/main/java/lems/cowshed/service/regular/event/RegularEventService.java
@@ -3,8 +3,6 @@ package lems.cowshed.service.regular.event;
 import lems.cowshed.api.controller.dto.regular.event.request.RegularEventEditRequest;
 import lems.cowshed.api.controller.dto.regular.event.request.RegularEventSaveRequest;
 import lems.cowshed.api.controller.dto.regular.event.response.RegularEventSimpleInfo;
-import lems.cowshed.api.controller.dto.regular.event.response.RegularParticipantsInfo;
-import lems.cowshed.api.controller.dto.regular.event.response.RegularParticipantDetails;
 import lems.cowshed.domain.event.Event;
 import lems.cowshed.domain.event.EventRepository;
 import lems.cowshed.domain.regular.event.RegularEvent;
@@ -12,30 +10,25 @@ import lems.cowshed.domain.regular.event.RegularEventEditCommand;
 import lems.cowshed.domain.regular.event.RegularEventRepository;
 import lems.cowshed.domain.regular.event.participation.RegularEventParticipation;
 import lems.cowshed.domain.regular.event.participation.RegularEventParticipationRepository;
-import lems.cowshed.domain.user.User;
-import lems.cowshed.domain.user.UserRepository;
-import lems.cowshed.exception.BusinessException;
 import lems.cowshed.exception.NotFoundException;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
 import static lems.cowshed.exception.Message.*;
 import static lems.cowshed.exception.Reason.*;
 
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Service
 public class RegularEventService {
 
-    private final UserRepository userRepository;
     private final EventRepository eventRepository;
     private final RegularEventRepository regularEventRepository;
     private final RegularEventParticipationRepository regularEventParticipationRepository;
 
+    @Transactional
     public Long saveRegularEvent(RegularEventSaveRequest request, Long eventId, Long userId) {
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new NotFoundException(EVENT_ID, EVENT_NOT_FOUND));
@@ -55,30 +48,7 @@ public class RegularEventService {
         return RegularEventSimpleInfo.from(regularEvent);
     }
 
-    public void saveParticipation(Long regularId, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException(USER_ID, USER_NOT_FOUND));
-
-        RegularEvent regularEvent = findRegularEvent(regularId);
-        regularEventParticipation(regularEvent);
-
-        regularEventParticipationRepository.save(RegularEventParticipation.of(user, regularEvent));
-    }
-
-    public RegularParticipantsInfo getRegularParticipants(Long regularId) {
-        List<Long> participantsUserIds = regularEventRepository.findParticipantsUserIdsByRegularId(regularId);
-        List<User> participants = userRepository.findByIdIn(participantsUserIds);
-        List<RegularParticipantDetails> details = convertDetails(participants);
-        return RegularParticipantsInfo.of(details, participantsUserIds.size());
-    }
-
-    public void deleteParticipation(Long participationId, Long userId) {
-        userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException(USER_ID, USER_NOT_FOUND));
-
-        regularEventParticipationRepository.deleteByIdAndUserId(participationId, userId);
-    }
-
+    @Transactional
     public void editRegularEvent(RegularEventEditRequest request, Long regularId) {
         RegularEvent regularEvent = regularEventRepository.findById(regularId)
                 .orElseThrow(() -> new NotFoundException(REGULAR_EVENT_ID, REGULAR_EVENT_NOT_FOUND));
@@ -95,24 +65,6 @@ public class RegularEventService {
                 .userId(userId)
                 .regularEvent(regularEvent)
                 .build();
-    }
-
-    private void regularEventParticipation(RegularEvent regularEvent) {
-        long participantCount = regularEventRepository.getParticipantCount(regularEvent.getId());
-        if(regularEvent.isNotPossibleParticipation(participantCount)){
-            throw new BusinessException(REGULAR_EVENT_PARTICIPATION, REGULAR_EVENT_NOT_POSSIBLE_PARTICIPATION);
-        }
-    }
-
-    private RegularEvent findRegularEvent(Long regularId) {
-        return regularEventRepository.findRegularEventWithLockById(regularId)
-                .orElseThrow(() -> new NotFoundException(REGULAR_EVENT_ID, REGULAR_EVENT_NOT_FOUND));
-    }
-
-    private List<RegularParticipantDetails> convertDetails(List<User> participants) {
-        return participants.stream()
-                .map(user -> RegularParticipantDetails.of(user.getUsername(), user.getMbti()))
-                .toList();
     }
 
     private RegularEventEditCommand createRegularEditCommand(RegularEventEditRequest request) {
